[change] 폴더 구조 변경
* router와 관련된 모듈들 router/ 폴더로 이동
* router.js -> router/router.js
* 각 router의 callback 함수들은 router/ 폴더 안에 생성
* ex)
  * router/register.js -> /register의 callback 함수
